backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
passlib[argon2]==1.7.*
python-jose[cryptography]==3.3.*
python-multipart==0.0.*
sqlalchemy[asyncio]==2.0.*
asyncpg==0.30.*
email-validator==2.2.*
pydantic-settings==2.7.*
python-dotenv==1.0.*
slowapi==0.1.*
